chatbot/studio/requirements.txt
python-dotenv
langchain-openai
langchain-community
faiss-cpu
langchain
langchain-core
langchain-text-splitters
typing-extensions
langgraph
selenium